requirements.txt
ollama
python-dotenv
openai
google-genai